Rename files to fit into DA structure
