+++ Lab1/src/java/DA_Schiper_Eggli_Sandoz_main.java
@@ -1,0 +1,4 @@
+package java;
+
+public class DA_Schiper_Eggli_Sandoz_main {
+}
